# frontend/src/types/openapi.yaml
...
openapi: 3.0.1
info:
  title: travelPlan API Test
  description: travelPlan의 api 테스트입니다.
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
security:
- bearerAuth: []
...
paths:
  "/trip/schedule/{scheduleId}":
    put:
      tags:
      - TripSchedule API
      operationId: updateSchedule
      parameters:
      - name: scheduleId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/TripScheduleUpdateReqDto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataTripScheduleResDto"
  "/trip/schedule/update":
    put:
      tags:
      - TripSchedule API
      summary: 나의 여행 일정 및 여행 정보 통합 수정
      operationId: updateTrip
      parameters:
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/TripUpdateReqDto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataTripUpdateResDto"
  "/trip/information/{tripInfoId}":
    put:
      tags:
      - trip-information-controller
      operationId: updateTripInformation
      parameters:
      - name: tripInfoId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/TripInformationUpdateReqDto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataTripInformationResDto"
...
  "/member/{id}":
    put:
      tags:
      - Member API
      summary: 회원정보 수정
      operationId: updateMember
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/MemberUpdateRequestDto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/MemberResponseDto"
    delete:
      tags:
      - Member API
      summary: 회원 삭제
      operationId: deleteMember
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
...
  "/trip/schedule":
    get:
      tags:
      - TripSchedule API
      operationId: getAllSchedules
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataListTripScheduleResDto"
    post:
      tags:
      - TripSchedule API
      summary: 나의 여행 일정 등록
      operationId: createSchedule
      parameters:
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/TripScheduleReqDto"
        required: true
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataTripScheduleResDto"
...
  "/trip/schedule/my-schedules":
    get:
      tags:
      - TripSchedule API
      summary: 나의 여행 일정 전체 조회
      operationId: getMySchedules
      parameters:
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataListTripScheduleResDto"
  "/trip/schedule/member/{memberId}":
    get:
      tags:
      - TripSchedule API
      operationId: getSchedulesByMember
      parameters:
      - name: memberId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataListTripScheduleResDto"
...
  "/trip/schedule/my-schedules/{scheduleId}":
    delete:
      tags:
      - TripSchedule API
      summary: 나의 여행 일정 삭제
      operationId: deleteSchedule
      parameters:
      - name: scheduleId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: Authorization
        in: header
        required: false
        schema:
          type: string
      responses:
        '200':
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/RsDataVoid"
...
components:
  schemas:
    TripScheduleUpdateReqDto:
      required:
      - description
      - endDate
      - startDate
      - title
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
    RsDataTripScheduleResDto:
      type: object
      properties:
        code:
          type: string
        msg:
          type: string
        data:
          "$ref": "#/components/schemas/TripScheduleResDto"
    TripInformationResDto:
      type: object
      properties:
        placeId:
          type: integer
          format: int64
        visitTime:
          type: string
          format: date-time
        duration:
          type: integer
          format: int32
        transportation:
          type: string
          enum:
          - WALK
          - BUS
          - SUBWAY
          - CAR
          - TAXI
          - ETC
        cost:
          type: integer
          format: int32
        notes:
          type: string
        priority:
          type: integer
          format: int32
        visited:
          type: boolean
    TripScheduleResDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        memberName:
          type: string
        title:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        tripInformations:
          type: array
          items:
            "$ref": "#/components/schemas/TripInformationResDto"
    TripInformationUpdateReqDto:
      required:
      - tripInformationId
      type: object
      properties:
        tripInformationId:
          type: integer
          format: int64
        placeId:
          type: integer
          format: int64
        visitTime:
          type: string
          format: date-time
        duration:
          type: integer
          format: int32
        transportation:
          type: string
          enum:
          - WALK
          - BUS
          - SUBWAY
          - CAR
          - TAXI
          - ETC
        cost:
          type: integer
          format: int32
        notes:
          type: string
        isVisited:
          type: boolean
    TripUpdateReqDto:
      required:
      - tripScheduleId
      type: object
      properties:
        tripScheduleId:
          type: integer
          format: int64
        scheduleUpdate:
          "$ref": "#/components/schemas/TripScheduleUpdateReqDto"
        tripInformationUpdates:
          type: array
          items:
            "$ref": "#/components/schemas/TripInformationUpdateReqDto"
    RsDataTripUpdateResDto:
      type: object
      properties:
        code:
          type: string
        msg:
          type: string
        data:
          "$ref": "#/components/schemas/TripUpdateResDto"
    TripUpdateResDto:
      type: object
      properties:
        updatedSchedule:
          "$ref": "#/components/schemas/TripScheduleResDto"
        updatedTripInformations:
          type: array
          items:
            "$ref": "#/components/schemas/TripInformationResDto"
    RsDataTripInformationResDto:
      type: object
      properties:
        code:
          type: string
        msg:
          type: string
        data:
          "$ref": "#/components/schemas/TripInformationResDto"
...
    MemberUpdateRequestDto:
      type: object
      properties:
        nickname:
          type: string
        email:
          type: string
        password:
          type: string
        profileImage:
          type: string
        gender:
          type: string
          enum:
          - MALE
          - FEMALE
        ageRange:
          type: string
          enum:
          - TEENS
          - TWENTIES
          - THIRTIES
          - FORTIES_PLUS
        travelStyle:
          type: string
          enum:
          - TOURISM
          - RELAXATION
          - SHOPPING
        aboutMe:
          type: string
    MemberResponseDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        nickname:
          type: string
        profileImage:
          type: string
        gender:
          type: string
          enum:
          - MALE
          - FEMALE
        ageRange:
          type: string
          enum:
          - TEENS
          - TWENTIES
          - THIRTIES
          - FORTIES_PLUS
        travelStyle:
          type: string
          enum:
          - TOURISM
          - RELAXATION
          - SHOPPING
        aboutMe:
          type: string
        rating:
          type: number
          format: double
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        authority:
          type: string
...
    TripInformationReqDto:
      required:
      - placeId
      type: object
      properties:
        placeId:
          type: integer
          format: int64
        visitTime:
          type: string
          format: date-time
        duration:
          type: integer
          format: int32
        transportation:
          type: string
          enum:
          - WALK
          - BUS
          - SUBWAY
          - CAR
          - TAXI
          - ETC
        cost:
          type: integer
          format: int32
        notes:
          type: string
        priority:
          type: integer
          format: int32
        visited:
          type: boolean
    TripScheduleReqDto:
      type: object
      properties:
        memberId:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        tripInformations:
          type: array
          items:
            "$ref": "#/components/schemas/TripInformationReqDto"
...
    RsDataListTripScheduleResDto:
      type: object
      properties:
        code:
          type: string
        msg:
          type: string
        data:
          type: array
          items:
            "$ref": "#/components/schemas/TripScheduleResDto"
...
    RsDataVoid:
      type: object
      properties:
        code:
          type: string
        msg:
          type: string
        data:
          type: object
  securitySchemes:
    bearerAuth:
      type: http
      name: Authorization
      in: header
      scheme: bearer
      bearerFormat: JWT